# Write index along the data writes instead of a dump on a close